requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.25.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
requests==2.31.0

# ML/AI dependencies
numpy==1.24.3
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
chromadb==0.4.18
openai==1.3.5
google-generativeai==0.8.3
anthropics==0.3.11

# Database and storage
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
sqlite3

# CLI and UI dependencies
rich==13.7.0
textual==0.45.1
click==8.1.7
typer==0.9.0
colorama==0.4.6

# System and monitoring
psutil==5.9.6
prometheus-client==0.18.0
watchdog==3.0.0

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-mock==3.12.0
pytest-benchmark==4.0.0
pytest-timeout==2.1.0
pytest-html==4.1.1
factory-boy==3.3.0
faker==20.1.0
testcontainers==3.7.1

# Code quality and linting
black==23.11.0
isort==5.12.0
mypy==1.7.0
flake8==6.1.0
pylint==3.0.3
bandit[toml]==1.7.5
safety==2.3.5
pre-commit==3.5.0

# Performance testing
locust==2.17.0
memory-profiler==0.61.0
line-profiler==4.1.1

# API and networking
aiohttp==3.9.1
typing-extensions==4.8.0
jsonschema==4.19.2
pydantic-core==2.14.5

# Utilities
python-dotenv==1.0.0
tomlkit==0.12.3
ruamel.yaml==0.18.5
markdown==3.5.1
jinja2-time==0.2.0